Change `to show` to `to display`.
`to show` is already used by News. There is nowhere else in the current mainline where the same `to *` phrase is used in different contexts; changing to `to display` maintains that invariant. It may also make it slightly more clear that the goto of a node is followed whether it's displayed or not.